+++ TornSharp/ApiModels/UserModels/Display/Display.cs
@@ -23,5 +23,5 @@ public class Display : JsonModel
     public string Type { get; set; }
 
     [JsonPropertyName("UID")]
-    public int? Uid { get; set; }
+    public long? Uid { get; set; }
 }
